Add computer choice logic
Create the getComputerChoice function that randomly
returns “rock”, “paper” or “scissors”.

To achive this functionality generate a random number
between 1 and 3 inclusively,use the
math.random builtin function(which returns a random number
that’s greater than or equal to 0 and less than 1)
with the math.floor function and add 1 to make
the random range between 1 and 3 inclusively.

Finally return the string according to the random number.