Add the logic to play multiple rounds
Create the playGame function that calls playRound function
to play multiple rounds(depends on the GAME_ROUNDS constant),
keeps track of the scores and declares a winner at the end.

Modify the playRound function so that it return a string
value to indicate the result of the round instead of directly
updating the scores variables.