Add ui logic and rework some functions due to that
Remove the logic that plays exactly five rounds.
Create three buttons, one for each selection. Add an event listener to
the buttons that call the playRound function with the correct
playerSelection every time a button is clicked.
Add a div for displaying results and change all the console.logs
into DOM methods.
Display the running score, and announce a winner of the game once
one player reaches 5 points.